Update Bug reports (Jira)
Bug: Error When Leaving the Login Fields Empty (https://www.saucedemo.com/)

Bug ID: LT-1
Description: This bug occurs when the user leaves the username or password fields empty on the login page. The system only shows an error for the "username" field, but does not show an error for the "password" field.

Steps to Reproduce:

Go to the login page.
Leave the username or password field empty.
Click the login button.
Expected Behavior: If either the username or password field is empty, the system should display an error message like "Please fill in the required fields" or "Username and Password cannot be empty."

Actual Behavior: When the username and password fields are empty, the system only shows an error for the "username" field, not for the "password" field.

Status: Open

Bug: Error in the Change Password Section (orangehrmlive.com)

Bug ID: LT-2
Description: This bug occurs in the change password functionality. After the user enters the "username" field, a link should be sent to the email, but there is no field to enter the email.

Steps to Reproduce:

Go to the Change Password section.
Enter the username.
Click the "Reset password" button.
Expected Behavior: There should be a field to enter the email.

Actual Behavior: There is no field to enter the email.

Status: Open